requirements.txt
selenium==2.41.0
virtualenv==1.11.4
coverage==3.6
dj-static==0.0.5
uwsgi=2.0.3
django-tinymce=1.5.2
django-data-migration=0.2.0
Pillow=2.4.0